.history/requirements_20251112131834.txt
Flask>=2.3.3
Flask-SQLAlchemy>=3.0.5
Flask-Caching>=2.0.2
Pillow>=10.0.0
pillow-heif>=0.14.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
redis>=4.5.4
gunicorn>=21.2.0
